bert_ner/requirements.txt
absl-py @ file:///opt/conda/conda-bld/absl-py_1639803114343/work
aiohttp @ file:///tmp/build/80754af9/aiohttp_1646806365504/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
astor==0.8.1
asttokens==2.1.0
astunparse==1.6.3
async-timeout @ file:///tmp/build/80754af9/async-timeout_1637851218186/work
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
backcall==0.2.0
beautifulsoup4==4.11.1
blinker==1.4
boto3==1.15.18
botocore==1.18.18
brotlipy==0.7.0
bs4==0.0.1
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
certifi==2022.9.24
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///tmp/build/80754af9/click_1646056590078/work
cryptography @ file:///tmp/build/80754af9/cryptography_1652101588893/work
Cython==0.29.32
debugpy==1.6.3
decorator==5.1.1
elasticsearch==7.17.7
elasticsearch-dsl==7.4.0
entrypoints==0.4
executing==1.2.0
filelock==3.8.0
Flask==2.2.2
flatbuffers==1.12
frozenlist @ file:///tmp/build/80754af9/frozenlist_1637767113340/work
fuzzywuzzy==0.18.0
gast==0.4.0
gluonnlp==0.10.0
google-auth @ file:///opt/conda/conda-bld/google-auth_1646735974934/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1617120569401/work
google-pasta @ file:///Users/ktietz/demo/mc3/conda-bld/google-pasta_1630577991354/work
graphviz==0.8.4
grpcio==1.34.1
h5py==3.1.0
haversine==2.7.0
huggingface-hub==0.0.12
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1648544546694/work
install==1.3.5
ipykernel==6.17.0
ipython==8.6.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
JPype1==1.4.1
jupyter_client==7.4.4
jupyter_core==4.11.2
keras==2.7.0
keras-crf==0.3.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
keras-self-attention==0.51.0
koalanlp==2.1.7
kobert @ git+https://git@github.com/SKTBrain/KoBERT.git@47a69af87928fc24e20f571fe10c3cc9dd9af9a3
kobert-tokenizer @ git+https://github.com/SKTBrain/KoBERT.git@47a69af87928fc24e20f571fe10c3cc9dd9af9a3#subdirectory=kobert_hf
konlpy==0.6.0
kss==2.5.1
libclang==14.0.6
lxml==4.9.1
Markdown @ file:///tmp/build/80754af9/markdown_1614363852612/work
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mecab-python===0.996-ko-0.9.2
mistletoe==0.9.0
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
multidict @ file:///opt/conda/conda-bld/multidict_1640703752579/work
mxnet==1.7.0.post2
nest-asyncio==1.5.6
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1626681925468/work
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.1.0
onnxruntime==1.8.0
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
packaging==21.3
pandas==1.4.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
plotly==5.11.0
prompt-toolkit==3.0.32
protobuf==3.20.1
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py-hanspell @ git+https://github.com/ssut/py-hanspell.git@8e993cf46f97f9d665c15633a0fc78ac1b727713
py4j==0.10.9.7
pyahocorasick==1.4.4
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.13.0
PyJWT @ file:///opt/conda/conda-bld/pyjwt_1657544592787/work
pykospacing @ git+https://github.com/haven-jeon/PyKoSpacing.git@e1979f7eeb4fcf1a456521152713482517fba627
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
pyzmq==24.0.1
regex==2022.10.31
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
s3transfer==0.3.7
sacremoses==0.0.53
scikit-learn==1.1.3
scipy @ file:///tmp/build/80754af9/scipy_1641555004408/work
sentencepiece==0.1.96
seqeval==1.2.2
silence-tensorflow==1.2.1
six==1.15.0
soupsieve==2.3.2.post1
stack-data==0.6.0
support-developer==1.0.5
tenacity==8.1.0
tensorboard @ file:///home/builder/tkoch/workspace/tensorflow/tensorboard_1658918251247/work/tensorboard-2.8.0-py3-none-any.whl
tensorboard-data-server @ file:///tmp/build/80754af9/tensorboard-data-server_1633034767313/work/tensorboard_data_server-0.6.0-py3-none-manylinux2010_x86_64.whl
tensorboard-plugin-wit @ file:///home/builder/tkoch/workspace/tensorflow/tensorboard-plugin-wit_1658918494740/work/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
tensorflow-addons==0.18.0
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==1.1.0
tf-crf-layer==0.2.5
tf2crf==0.1.33
threadpoolctl==3.1.0
tokenizer==3.4.2
tokenizer-tools==0.46.1
tokenizers==0.10.3
torch==1.10.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
transformers==4.8.1
typeguard==2.13.3
typing-extensions==3.7.4.3
urllib3==1.25.11
wcwidth==0.2.5
Werkzeug==2.2.2
wikiextractor==3.0.6
wrapt==1.12.1
yarl @ file:///tmp/build/80754af9/yarl_1606939947528/work
zipp @ file:///opt/conda/conda-bld/zipp_1652341764480/work
